finish the last of the properties articles
