+++ docs/csharp/programming-guide/classes-and-structs/auto-implemented-properties.md
@@ -1,23 +1,22 @@
 ---
 title: "Auto-Implemented Properties - C# Programming Guide"
 description: For an auto-implemented property in C#, the compiler creates a private, anonymous backing field accessed only through get and set accessors of the property.
-ms.date: 01/31/2020
+ms.date: 07/29/2022
 f1_keywords:
   - "propertyInitializer_CSharpKeyword"
 helpviewer_keywords:
   - "auto-implemented properties [C#]"
   - "properties [C#], auto-implemented"
-ms.assetid: aa55fa97-ccec-431f-b5e9-5ac789fd32b7
 ---
 # Auto-Implemented Properties (C# Programming Guide)
 
-In C# 3.0 and later, auto-implemented properties make property-declaration more concise when no additional logic is required in the property accessors. They also enable client code to create objects. When you declare a property as shown in the following example, the compiler creates a private, anonymous backing field that can only be accessed through the property's `get` and `set` accessors. In C# 9 and later, `init` accessors can also be declared as auto-implemented properties.
+Auto-implemented properties make property-declaration more concise when no additional logic is required in the property accessors. They also enable client code to create objects. When you declare a property as shown in the following example, the compiler creates a private, anonymous backing field that can only be accessed through the property's `get` and `set` accessors. In C# 9 and later, `init` accessors can also be declared as auto-implemented properties.
 
 ## Example
 
 The following example shows a simple class that has some auto-implemented properties:
 
-[!code-csharp[csProgGuideLINQ#28](~/samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideLINQ/CS/csRef30LangFeatures_2.cs#28)]
+:::code language="csharp" source="./snippets/properties/AutoImplemented.cs" id="Snippet28":::
 
 You can't declare auto-implemented properties in interfaces. Auto-implemented properties declare a private instance backing field, and interfaces may not declare instance fields. Declaring a property in an interface without defining a body declares a property with accessors that must be implemented by each type that implements that interface.
